requirements-dev.txt
torch~=2.4.0+cpu
fastonn~=0.1.1
wandb~=0.17.5
torchvision~=0.19.0+cpu
tqdm~=4.66.4
numpy~=1.26.4
scikit-image~=0.24.0
scikit-learn~=1.4.2
pillow~=10.4.0
pytest~=8.3.4